# refactor(BaseDocument)!: improved `get`, `set` and `extend` methods