--- a/frappe/model/meta.py
+++ b/frappe/model/meta.py
@@ -30,7 +30,7 @@ from frappe.model import (
 	optional_fields,
 	table_fields,
 )
-from frappe.model.base_document import BaseDocument
+from frappe.model.base_document import DOCTYPE_TABLE_FIELDS, BaseDocument
 from frappe.model.document import Document
 from frappe.model.workflow import get_workflow_name
 from frappe.modules import load_doctype_module
@@ -640,14 +640,6 @@ class Meta(Document):
 	def is_nested_set(self):
 		return self.has_field("lft") and self.has_field("rgt")
 
-
-DOCTYPE_TABLE_FIELDS = [
-	frappe._dict({"fieldname": "fields", "options": "DocField"}),
-	frappe._dict({"fieldname": "permissions", "options": "DocPerm"}),
-	frappe._dict({"fieldname": "actions", "options": "DocType Action"}),
-	frappe._dict({"fieldname": "links", "options": "DocType Link"}),
-	frappe._dict({"fieldname": "states", "options": "DocType State"}),
-]
 
 #######