This PR continues where https://github.com/frappe/frappe/pull/16477 left off.

The only breaking change is that if someone does `doc.my_table = None` and then does `doc.get("my_table")` they will get `None`, not `[]`

`doc.set("my_table", None)` will set `doc.my_table` to an empty list.

## Additional changes
- `doc.set` will now call `doc.extend` only for table fields. This means one can now set doc properties to a list without requiring those properties to be child tables.
- Init child tables to an empty list for all falsy values like `None` in `doc.set`
- Allow all iterable objects in `doc.extend`, instead of just list
- Remove no longer useful code from `doc.append`
- Remove unnecessary if condition in `doc._init_child`
- Introduce new constant `TABLE_DOCTYPES_FOR_DOCTYPE` to avoid hardcoding
- Change definition of `DOCTYPES_FOR_DOCTYPE` to avoid repetition
- `ignore_in_setter` => `_reserved_keywords` (key not released for a stable version yet)
- Improved DocType validation with more fields in `_reserved_keywords`
- Convert `DOCTYPES_FOR_DOCTYPE` and `_reserved_keywords` to sets. Sets are relatively faster for lookup, since they implement hash-based indexing.
- Added new reserved property `_table_fieldnames`: used in `doc.set`
- (minor) reduce `getattr` in `meta` property
- (minor) cleaner `doc.get`

---

### `doc.get`

#### Before

```python

In [3]: %timeit doc.get("key")
1.15 µs ± 1.26 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

```

#### After (83% faster)

```python
In [3]: %timeit doc.get("key")
192 ns ± 0.158 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

### `doc.set("key", None)`

#### Before

```python
In [4]: %timeit doc.set("key", None)
241 ns ± 0.542 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: %timeit doc.set("key", None)
243 ns ± 0.108 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [6]: %timeit doc.set("key", None)
239 ns ± 0.178 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)


```

#### After (surprisingly slightly faster - probably because we used Python sets)

```python
In [2]: %timeit doc.set("key", None)
236 ns ± 5.65 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
https://github.com/frappe/erpnext/runs/5870570026?check_suite_focus=true
In [2]: %timeit doc.set("key", None)
236 ns ± 0.231 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [3]: %timeit doc.set("key", None)
236 ns ± 0.262 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit doc.set("key", None)
236 ns ± 0.618 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

---

[Tested with ERPNext](https://github.com/frappe/erpnext/actions/runs/2109682414) 😄